split the tree file
